Separate partner type unit tests into own class
